+++ et-driver/et-soc1-pcie_loopback.c
@@ -33,7 +33,7 @@
 #include "et_io.h"
 #include "et_ioctl.h"
 #include "et_pci_dev.h"
-#include "et_sysfs_stats.h"
+#include "et_sysfs.h"
 #include "et_vma.h"
 #include "et_vqueue.h"
 
@@ -1058,23 +1058,22 @@ int et_mgmt_dev_init(struct et_pci_dev *et_dev,
 		goto error_unlock_init_mutex;
 	}
 
+	// SysFs error statistics initialization
+	rv = et_sysfs_add_group(et_dev, ET_SYSFS_GID_ERR_STATS);
+	if (rv) {
+		dev_err(&et_dev->pdev->dev,
+			"Mgmt: et_sysfs_add_group() failed, group_id: %d\n",
+			ET_SYSFS_GID_ERR_STATS);
+		goto error_free_vq_buffer;
+	}
 	et_err_stats_init(&et_dev->mgmt.err_stats);
 
 	// VQs initialization
 	rv = et_vqueue_init_all(et_dev, true /* mgmt_dev */);
 	if (rv) {
 		dev_err(&et_dev->pdev->dev,
-			"Mgmt: VQs initialization failed\n");
-		goto error_free_vq_buffer;
-	}
-
-	// SysFs statistics initialization
-	rv = et_sysfs_stats_init(et_dev, true /* mgmt_dev */);
-	if (rv) {
-		dev_err(&et_dev->pdev->dev,
-			"Mgmt: Failed to init sysfs files, error %d\n",
-			-rv);
-		goto error_vqueue_destroy_all;
+			"Mgmt: et_vqueue_init_all() failed!\n");
+		goto error_sysfs_remove_group;
 	}
 
 	if (miscdev_create && !et_dev->mgmt.miscdev_created) {
@@ -1089,7 +1088,7 @@ int et_mgmt_dev_init(struct et_pci_dev *et_dev,
 		if (rv) {
 			dev_err(&et_dev->pdev->dev,
 				"Mgmt: misc register failed\n");
-			goto error_sysfs_stats_remove;
+			goto error_vqueue_destroy_all;
 		}
 		et_dev->mgmt.miscdev_created = true;
 	}
@@ -1099,11 +1098,11 @@ int et_mgmt_dev_init(struct et_pci_dev *et_dev,
 
 	return rv;
 
-error_sysfs_stats_remove:
-	et_sysfs_stats_remove(et_dev, true /* mgmt_dev */);
-
 error_vqueue_destroy_all:
 	et_vqueue_destroy_all(et_dev, true /* mgmt_dev */);
+
+error_sysfs_remove_group:
+	et_sysfs_remove_group(et_dev, ET_SYSFS_GID_ERR_STATS);
 
 error_free_vq_buffer:
 	kfree((void __force *)et_dev->mgmt
@@ -1129,8 +1128,8 @@ void et_mgmt_dev_destroy(struct et_pci_dev *et_dev, bool miscdev_destroy)
 	if (!et_dev->mgmt.is_initialized)
 		goto unlock_init_mutex;
 
-	et_sysfs_stats_remove(et_dev, true /* mgmt_dev */);
 	et_vqueue_destroy_all(et_dev, true /* mgmt_dev */);
+	et_sysfs_remove_group(et_dev, ET_SYSFS_GID_ERR_STATS);
 
 	kfree((void __force *)et_dev->mgmt
 		      .regions[MGMT_MEM_REGION_TYPE_VQ_BUFFER]
@@ -1207,23 +1206,22 @@ int et_ops_dev_init(struct et_pci_dev *et_dev,
 	region->size = 0x2FF000000ULL;
 	region->mapped_baseaddr = NULL;
 
+	// SysFs memory statistics initialization
+	rv = et_sysfs_add_group(et_dev, ET_SYSFS_GID_MEM_STATS);
+	if (rv) {
+		dev_err(&et_dev->pdev->dev,
+			"Ops: et_sysfs_add_group() failed, group_id: %d\n",
+			ET_SYSFS_GID_MEM_STATS);
+		goto error_free_vq_buffer;
+	}
 	et_mem_stats_init(&et_dev->ops.mem_stats);
 
 	// VQs initialization
 	rv = et_vqueue_init_all(et_dev, false /* ops_dev */);
 	if (rv) {
 		dev_err(&et_dev->pdev->dev,
-			"Ops device VQs initialization failed\n");
-		goto error_free_vq_buffer;
-	}
-
-	// SysFs statistics initialization
-	rv = et_sysfs_stats_init(et_dev, false /* ops_dev */);
-	if (rv) {
-		dev_err(&et_dev->pdev->dev,
-			"Ops: Failed to init sysfs files, error %d\n",
-			-rv);
-		goto error_vqueue_destroy_all;
+			"Ops: et_vqueue_init_all() failed!\n");
+		goto error_sysfs_remove_group;
 	}
 
 	if (miscdev_create && !et_dev->ops.miscdev_created) {
@@ -1238,7 +1236,7 @@ int et_ops_dev_init(struct et_pci_dev *et_dev,
 		if (rv) {
 			dev_err(&et_dev->pdev->dev,
 				"misc ops register failed\n");
-			goto error_sysfs_stats_remove;
+			goto error_vqueue_destroy_all;
 		}
 		et_dev->ops.miscdev_created = true;
 	}
@@ -1248,11 +1246,11 @@ int et_ops_dev_init(struct et_pci_dev *et_dev,
 
 	return rv;
 
-error_sysfs_stats_remove:
-	et_sysfs_stats_remove(et_dev, false /* ops_dev */);
-
 error_vqueue_destroy_all:
 	et_vqueue_destroy_all(et_dev, false /* ops_dev */);
+
+error_sysfs_remove_group:
+	et_sysfs_remove_group(et_dev, ET_SYSFS_GID_MEM_STATS);
 
 error_free_vq_buffer:
 	kfree((void __force *)et_dev->ops.regions[OPS_MEM_REGION_TYPE_VQ_BUFFER]
@@ -1277,8 +1275,8 @@ void et_ops_dev_destroy(struct et_pci_dev *et_dev, bool miscdev_destroy)
 	if (!et_dev->ops.is_initialized)
 		goto unlock_init_mutex;
 
-	et_sysfs_stats_remove(et_dev, false /* ops_dev */);
 	et_vqueue_destroy_all(et_dev, false /* ops_dev */);
+	et_sysfs_remove_group(et_dev, ET_SYSFS_GID_MEM_STATS);
 
 	kfree((void __force *)et_dev->ops.regions[OPS_MEM_REGION_TYPE_VQ_BUFFER]
 		      .mapped_baseaddr);
@@ -1376,8 +1374,7 @@ static void uninit_et_pci_dev(struct et_pci_dev *et_dev, bool miscdev_destroy)
 static void et_reset_isr_work(struct work_struct *work)
 {
 	int rv = 0;
-	int retry = 0;
-	int hit_count = 0;
+	unsigned int wait_ms = 100, time_ms, uptime_ms;
 	struct et_pci_dev *et_dev =
 		container_of(work, struct et_pci_dev, isr_work);
 
@@ -1395,27 +1392,28 @@ static void et_reset_isr_work(struct work_struct *work)
 	dev_dbg(&et_dev->pdev->dev, "Waiting for PCIe link to settle...\n");
 	// After reset is triggered, the device goes down after some time and
 	// then gets up again. To detect that the link is stable, the device
-	// should be present for consecutive 300ms
-	for (retry = 0; retry < 20 && hit_count < 3; retry++) {
+	// should be present for `pcilink_max_estim_downtime_ms`
+	for (time_ms = 0, uptime_ms = 0;
+	     time_ms < et_dev->reset_cfg.pcilink_discovery_timeout_ms &&
+	     uptime_ms < et_dev->reset_cfg.pcilink_max_estim_downtime_ms;
+	     time_ms += wait_ms) {
 		if (pci_device_is_present(et_dev->pdev))
-			hit_count++;
+			uptime_ms += wait_ms;
 		else
-			hit_count = 0;
-		msleep(100);
+			uptime_ms = 0;
+		msleep(wait_ms);
 	}
 
-	if (hit_count < 3) {
-		// TODO: SW-15433: Add handling for device when it doesn't come
-		// up with above retries
+	if (uptime_ms < et_dev->reset_cfg.pcilink_max_estim_downtime_ms) {
 		dev_err(&et_dev->pdev->dev,
-			"Unable to detect the device on bus after reset!");
+			"Unable to detect the device on bus!");
 		goto exit_reset;
 	}
 
 	rv = pci_load_saved_state(et_dev->pdev, et_dev->pstate);
 	if (rv) {
 		dev_warn(&et_dev->pdev->dev,
-			 "Failed to load PCI state, err: %d\n",
+			 "Failed to load PCI state, errno: %d\n",
 			 -rv);
 	} else {
 		pci_restore_state(et_dev->pdev);
@@ -1463,6 +1461,15 @@ static int esperanto_pcie_probe(struct pci_dev *pdev,
 		goto error_free_saved_state;
 	}
 
+	rv = et_sysfs_add_group(et_dev, ET_SYSFS_GID_SOC_RESET);
+	if (rv) {
+		dev_err(&et_dev->pdev->dev,
+			"et_sysfs_add_group() failed, group_id: %d\n",
+			ET_SYSFS_GID_SOC_RESET);
+		goto error_uninit_et_pci_dev;
+	}
+	et_soc_reset_cfg_init(&et_dev->reset_cfg);
+
 	et_dev->reset_workqueue = alloc_workqueue("%s:et%d_rstwq",
 						  WQ_MEM_RECLAIM | WQ_UNBOUND,
 						  1,
@@ -1471,11 +1478,14 @@ static int esperanto_pcie_probe(struct pci_dev *pdev,
 	if (!et_dev->reset_workqueue) {
 		dev_err(&pdev->dev, "Mgmt device initialization failed\n");
 		rv = -ENOMEM;
-		goto error_uninit_et_pci_dev;
+		goto error_sysfs_remove_group;
 	}
 	INIT_WORK(&et_dev->isr_work, et_reset_isr_work);
 
 	return rv;
+
+error_sysfs_remove_group:
+	et_sysfs_remove_group(et_dev, ET_SYSFS_GID_SOC_RESET);
 
 error_uninit_et_pci_dev:
 	uninit_et_pci_dev(et_dev, true);
@@ -1504,6 +1514,8 @@ static void esperanto_pcie_remove(struct pci_dev *pdev)
 	destroy_workqueue(et_dev->reset_workqueue);
 
 	uninit_et_pci_dev(et_dev, true);
+	et_sysfs_remove_files(et_dev);
+	et_sysfs_remove_groups(et_dev);
 	if (et_dev->pstate) {
 		kfree(et_dev->pstate);
 		et_dev->pstate = NULL;