+++ et-driver/et_vqueue_loopback.c
@@ -1039,6 +1039,7 @@ ssize_t et_vqueue_init_all(struct et_pci_dev *et_dev, bool is_mgmt)
 {
 	ssize_t rv;
 	struct et_vq_common *vq_common;
+	int vq_stats_gid;
 
 	if (is_mgmt) {
 		vq_common = &et_dev->mgmt.vq_data.vq_common;
@@ -1048,6 +1049,8 @@ ssize_t et_vqueue_init_all(struct et_pci_dev *et_dev, bool is_mgmt)
 		vq_common->hp_sq_size = 0;
 		vq_common->cq_count = et_dev->mgmt.dir_vq.cq_count;
 		vq_common->cq_size = et_dev->mgmt.dir_vq.cq_size;
+
+		vq_stats_gid = ET_SYSFS_GID_MGMT_VQ_STATS;
 	} else {
 		vq_common = &et_dev->ops.vq_data.vq_common;
 		vq_common->sq_count = et_dev->ops.dir_vq.sq_count;
@@ -1056,6 +1059,8 @@ ssize_t et_vqueue_init_all(struct et_pci_dev *et_dev, bool is_mgmt)
 		vq_common->hp_sq_size = et_dev->ops.dir_vq.hp_sq_size;
 		vq_common->cq_count = et_dev->ops.dir_vq.cq_count;
 		vq_common->cq_size = et_dev->ops.dir_vq.cq_size;
+
+		vq_stats_gid = ET_SYSFS_GID_OPS_VQ_STATS;
 	}
 
 	bitmap_zero(vq_common->sq_bitmap, ET_MAX_QUEUES);
@@ -1065,9 +1070,13 @@ ssize_t et_vqueue_init_all(struct et_pci_dev *et_dev, bool is_mgmt)
 	init_waitqueue_head(&vq_common->waitqueue);
 	vq_common->pdev = et_dev->pdev;
 
-	rv = et_high_priority_squeue_init_all(et_dev, is_mgmt);
+	rv = et_sysfs_add_group(et_dev, vq_stats_gid);
 	if (rv)
 		return rv;
+
+	rv = et_high_priority_squeue_init_all(et_dev, is_mgmt);
+	if (rv)
+		goto error_sysfs_remove_group;
 
 	rv = et_squeue_init_all(et_dev, is_mgmt);
 	if (rv)
@@ -1084,6 +1093,9 @@ error_squeue_destroy_all:
 
 error_high_priority_squeue_destroy_all:
 	et_high_priority_squeue_destroy_all(et_dev, is_mgmt);
+
+error_sysfs_remove_group:
+	et_sysfs_remove_group(et_dev, vq_stats_gid);
 
 	return rv;
 }
@@ -1146,8 +1158,11 @@ static void et_cqueue_destroy_all(struct et_pci_dev *et_dev, bool is_mgmt)
 void et_vqueue_destroy_all(struct et_pci_dev *et_dev, bool is_mgmt)
 {
 	struct et_vq_data *vq_data;
+	int vq_stats_gid;
 
 	vq_data = is_mgmt ? &et_dev->mgmt.vq_data : &et_dev->ops.vq_data;
+	vq_stats_gid = is_mgmt ? ET_SYSFS_GID_MGMT_VQ_STATS :
+				 ET_SYSFS_GID_OPS_VQ_STATS;
 	wake_up_all(&vq_data->vq_common.waitqueue);
 	while (waitqueue_active(&vq_data->vq_common.waitqueue))
 		msleep(1);
@@ -1155,6 +1170,8 @@ void et_vqueue_destroy_all(struct et_pci_dev *et_dev, bool is_mgmt)
 	et_cqueue_destroy_all(et_dev, is_mgmt);
 	et_squeue_destroy_all(et_dev, is_mgmt);
 	et_high_priority_squeue_destroy_all(et_dev, is_mgmt);
+
+	et_sysfs_remove_group(et_dev, vq_stats_gid);
 
 	mutex_destroy(&vq_data->vq_common.sq_bitmap_mutex);
 	mutex_destroy(&vq_data->vq_common.cq_bitmap_mutex);